Ajout des méthodes dans la couche service

+++ assessment/src/main/java/com/mediscreen/assessment/proxies/PatientProxy.java
@@ -15,6 +15,6 @@ public interface PatientProxy {
     List<PatientDTO> getPatients();
 
     @GetMapping("/{id}")
-    PatientDTO getPatientById(@PathVariable("id") Integer id);
+    PatientDTO getPatientById(@PathVariable("id") Long id);
 
 }
